Fixed bugs and added timer
I have mad it so it doesn't crash when the respawn key is spammed - added time delay - and made it so if you respawn before the first checkpoint you go to the start line. I have added a HUD the displays a timer that counts up and your final time is displayed when you go through the finish line along with "FINISH!!" text.