Forced to fly through checkpoints to finish
The player now needs to fly through all checkpoints to finish. Checkpoints are counted at the start of the game, 1 is subtracted off this every time the player passes through a CP for the first time. When this = 0 the player can finish. This prevents players from skipping the course and prevents bug where game randomly finishes.