Added world border with sphere mask
I have added a world border to my maps that the player can't fly through - they bounce off. This border is invisible to the player when they are far away but when they get close, I used a sphere mask to make a translucent blue material to appear on the wall, showing the player where it is.